Reorganize pds message queue code out of db
